// apps/order/src/utilities/order.pb.ts
export interface CreateOrderDto {
  amount: number;
  userId: number;
  orderStatusId: number;
  orderItems: CreateOrderItemWithOrderDto[];
}
// ...
export interface PaginationDto {
  page: number;
  skip: number;
}
// ...
export interface CreateOrderItemWithOrderDto {
  amount: number;
  serviceName: string;
}